+++ routes/User.js
@@ -23,4 +23,4 @@ const UserSchema = new Schema({
   }
 })
 
-module.exports = User = mongoose.model('users', UserSchema)
+module.exports = User = mongoose.model('userpass', UserSchema)
