+++ routes/Users.js
@@ -2,9 +2,8 @@ const express = require('express')
 const users = express.Router()
 const cors = require('cors')
 const jwt = require('jsonwebtoken')
-const bcrypt = require('bcrypt')
 
-const User = require('../models/User')
+const User = require('./User')
 users.use(cors())
 
 process.env.SECRET_KEY = 'secret'
@@ -24,8 +23,8 @@ users.post('/register', (req, res) => {
   })
     .then(user => {
       if (!user) {
-        bcrypt.hash(req.body.password, 10, (err, hash) => {
-          userData.password = hash
+        
+          userData.password = "pass"
           User.create(userData)
             .then(user => {
               res.json({ status: user.email + 'Registered!' })
@@ -33,7 +32,7 @@ users.post('/register', (req, res) => {
             .catch(err => {
               res.send('error: ' + err)
             })
-        })
+       
       } else {
         res.json({ error: 'User already exists' })
       }
@@ -49,7 +48,6 @@ users.post('/login', (req, res) => {
   })
     .then(user => {
       if (user) {
-        if (bcrypt.compareSync(req.body.password, user.password)) {
           // Passwords match
           const payload = {
             _id: user._id,
@@ -61,10 +59,7 @@ users.post('/login', (req, res) => {
             expiresIn: 1440
           })
           res.send(token)
-        } else {
-          // Passwords don't match
-          res.json({ error: 'User does not exist' })
-        }
+        
       } else {
         res.json({ error: 'User does not exist' })
       }
