--- routes2/Users.js
@@ -1,90 +1,0 @@
-const express = require('express')
-const users = express.Router()
-const cors = require('cors')
-const jwt = require('jsonwebtoken')
-
-const User = require('./User')
-users.use(cors())
-
-process.env.SECRET_KEY = 'secret'
-
-users.post('/register', (req, res) => {
-  const today = new Date()
-  const userData = {
-    first_name: req.body.first_name,
-    last_name: req.body.last_name,
-    email: req.body.email,
-    password: req.body.password,
-    created: today
-  }
-
-  User.findOne({
-    email: req.body.email
-  })
-    .then(user => {
-      if (!user) {
-        
-          userData.password = "pass"
-          User.create(userData)
-            .then(user => {
-              res.json({ status: user.email + 'Registered!' })
-            })
-            .catch(err => {
-              res.send('error: ' + err)
-            })
-       
-      } else {
-        res.json({ error: 'User already exists' })
-      }
-    })
-    .catch(err => {
-      res.send('error: ' + err)
-    })
-})
-
-users.post('/login', (req, res) => {
-  User.findOne({
-    email: req.body.email
-  })
-    .then(user => {
-      if (user) {
-          // Passwords match
-          const payload = {
-            _id: user._id,
-            first_name: user.first_name,
-            last_name: user.last_name,
-            email: user.email
-          }
-          let token = jwt.sign(payload, process.env.SECRET_KEY, {
-            expiresIn: 1440
-          })
-          res.send(token)
-        
-      } else {
-        res.json({ error: 'User does not exist' })
-      }
-    })
-    .catch(err => {
-      res.send('error: ' + err)
-    })
-})
-
-users.get('/profile', (req, res) => {
-  var decoded = jwt.verify(req.headers['authorization'], process.env.SECRET_KEY)
-
-  User.findOne({
-    _id: decoded._id
-  })
-    .then(user => {
-      if (user) {
-        res.json(user)
-      } else {
-        res.send('User does not exist')
-      }
-    })
-    .catch(err => {
-      res.send('error: ' + err)
-    })
-})
-
-module.exports = users
