+++ server12.js
@@ -1,0 +1,30 @@
+var express = require('express')
+var cors = require('cors')
+var bodyParser = require('body-parser')
+var app = express()
+const mongoose = require('mongoose')
+var port = process.env.PORT || 3001
+
+app.use(bodyParser.json())
+app.use(cors())
+app.use(
+  bodyParser.urlencoded({
+    extended: false
+  })
+)
+const mongoURI = 'mongodb://localhost/reg'
+mongoose
+  .connect(
+    mongoURI,
+    { useNewUrlParser: true }
+  )
+  .then(() => console.log('MongoDB Connected'))
+  .catch(err => console.log(err))
+
+var Userspass = require('./routes/Users')
+
+app.use('/', Userspass)
+
+app.listen(port, function() {
+  console.log('Server is running on port: ' + port)
+})
